gpt: usize everything

+++ ocibootstrap/src/layout.rs
@@ -133,7 +133,7 @@ pub(crate) struct GptPartition {
     pub(crate) uuid: Uuid,
     pub(crate) name: Option<String>,
     pub(crate) mnt: PathBuf,
-    pub(crate) size_bytes: Option<u64>,
+    pub(crate) size_bytes: Option<usize>,
     pub(crate) fs: Filesystem,
     pub(crate) bootable: bool,
     pub(crate) platform_required: bool,
@@ -224,7 +224,7 @@ impl PartitionTable {
                     "com.github.mripard.ocibootstrap.partition.{part_name}.size_mb",
                 ))
                 .map(|size_str| {
-                    u64::from_str(size_str)
+                    usize::from_str(size_str)
                         .map(|size_mb| size_mb << 20)
                         .map_err(|_err| {
                             OciBootstrapError::Custom(format!(
